Edit pm-flash to support the flashing halium images
halium images are stored on images.plasma-mobile.org, it is notable that
current flash workflow is not set in stone and will be changes as soon
as how halium starts distributing images.

Most likely we will switch to recovery and OTA/channel based upgrades
which allows us to have proper multiple devices support and versioning
support as well.

This is just intended for testing while we sort out halium distribution
and OTA/channel based upgrades for plasma mobile.